requirements.txt
customtkinter>=5.2.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
gitpython>=3.1.40